Make the mod info header readable (#8)
* Make the mod info header readable

Unless I am misunderstanding something, there is no "modpack" here. This strange format makes it difficult for launchers to understand what is going on, thereby treating it as an arbitrary junk file.

* Is this it?

Going along with the suggested changes. It appears to do the trick, although I remain uncertain about the choice of format.

* Add in the needed variables

It was copied from static files, and I forgot about the variables that the data file uses in code.

* Tweak an omission